requirements.txt
pytest
selenium
webdriver_manager
allure-pytest
pytest-dependency
allure-pytest
